Rename base package, removes ant and adds gradle
