--- .classpath
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<classpath>
-	<classpathentry kind="src" path="src"/>
-	<classpathentry kind="src" path="test"/>
-	<classpathentry kind="con" path="org.eclipse.jdt.launching.JRE_CONTAINER/org.eclipse.jdt.internal.debug.ui.launcher.StandardVMType/JavaSE-1.6"/>
-	<classpathentry kind="lib" path="artifacts/test/hamcrest-all-1.1.jar"/>
-	<classpathentry kind="lib" path="artifacts/test/junit-4.8.1.jar"/>
-	<classpathentry kind="lib" path="artifacts/test/mockito-all-1.8.5.jar"/>
-	<classpathentry kind="output" path="bin"/>
-</classpath>
