+++ .gitignore
@@ -1,3 +1,13 @@
+.idea
+lib
 artifacts
+out
 bin
 build
+*.iml
+*.ipr
+*.iws
+.classpath
+.project
+.settings
+.gradle
