+++ build.gradle
@@ -1,0 +1,16 @@
+plugins {
+    id 'java-library'
+}
+
+dependencies {
+    testImplementation 'junit:junit:4.12'
+    testImplementation 'org.hamcrest:hamcrest-all:1.3'
+    testImplementation 'org.mockito:mockito-all:1.10.19'
+
+
+}
+
+repositories {
+    jcenter()
+    mavenCentral()
+}
