--- build.xml
@@ -1,35 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<project xmlns:ivy="antlib:org.apache.ivy.ant" default="build" name="coollection" basedir=".">
-	
-	<path id="lib">
-        <fileset dir="lib" includes="**/*.jar"/>
-		<fileset dir="artifacts/main" includes="**/*.jar"/>
-    </path>
-	
-	<path id="test">
-        <fileset dir="artifacts/test" includes="**/*.jar"/>
-    </path>
-	
-	<taskdef resource="org/apache/ivy/ant/antlib.xml" uri="antlib:org.apache.ivy.ant" classpathref="lib" />
-
-	<target name="resolve">
-        <ivy:retrieve pattern="artifacts/[conf]/[artifact]-[revision].[ext]" />
-    </target>
-	
-	<target name="clean">
-        <delete dir="build"/>
-    </target>
-
-    <target name="compile">
-        <mkdir dir="build/classes"/>
-        <javac srcdir="src" destdir="build/classes"><classpath refid="lib"/></javac>
-    </target>
-	
-	<target name="build" depends="resolve, clean, compile">
-		<delete file="build/coollection.jar"/>
-		<jar destfile="build/coollection.jar" basedir="build/classes">
-			<fileset dir="src"></fileset>
-		</jar>
-    </target>
-	
-</project>
