+++ gradle/wrapper/gradle-wrapper.properties
@@ -1,0 +1,6 @@
+#Mon Mar 05 23:49:33 BRT 2018
+distributionBase=GRADLE_USER_HOME
+distributionPath=wrapper/dists
+zipStoreBase=GRADLE_USER_HOME
+zipStorePath=wrapper/dists
+distributionUrl=https\://services.gradle.org/distributions/gradle-4.5.1-all.zip
