--- ivy.xml
@@ -1,12 +1,0 @@
-<ivy-module version="2.0">
-    <info organisation="com.wagnerandrade" module="belchior"/>
-    <configurations defaultconfmapping="main->default;test->default">
-      <conf name="main"/>
-      <conf name="test"/>
-    </configurations>
-    <dependencies>
-		<dependency conf="test" org="junit" name="junit" rev="4.8.1"/>
-        <dependency conf="test" org="org.mockito" name="mockito-all" rev="1.8.5"/>
-        <dependency conf="test" org="org.hamcrest" name="hamcrest-all" rev="1.1"/>
-    </dependencies>
-</ivy-module>
