--- src/com/wagnerandade/coollection/matcher/Matcher.java
@@ -1,7 +1,0 @@
-package com.wagnerandade.coollection.matcher;
-
-public interface Matcher {
-
-	public boolean match(Object value);
-
-}
