--- src/com/wagnerandade/coollection/query/order/Order.java
@@ -1,5 +1,0 @@
-package com.wagnerandade.coollection.query.order;
-
-public enum Order {
-	ASC, DESC
-}
