+++ src/main/java/org/simple/coollection/matcher/Matcher.java
@@ -1,0 +1,7 @@
+package org.simple.coollection.matcher;
+
+public interface Matcher {
+
+	public boolean match(Object value);
+
+}
