+++ src/main/java/org/simple/coollection/matcher/custom/Contains.java
@@ -1,0 +1,18 @@
+package org.simple.coollection.matcher.custom;
+
+import org.simple.coollection.matcher.Matcher;
+
+public class Contains implements Matcher {
+
+	private final String matcherValue;
+
+	public Contains(String matcherValue) {
+		this.matcherValue = matcherValue;
+	}
+
+	@Override
+	public boolean match(Object value) {
+		return String.valueOf(value).contains(matcherValue);
+	}
+
+}
