+++ src/main/java/org/simple/coollection/matcher/custom/EqualsIgnoreCase.java
@@ -1,0 +1,18 @@
+package org.simple.coollection.matcher.custom;
+
+import org.simple.coollection.matcher.Matcher;
+
+public class EqualsIgnoreCase implements Matcher {
+
+	private final String value;
+
+	public EqualsIgnoreCase(String value) {
+		this.value = value;
+	}
+
+	@Override
+	public boolean match(Object anotherValue) {
+		return (value).equalsIgnoreCase((String)anotherValue);
+	}
+
+}
