+++ src/main/java/org/simple/coollection/matcher/custom/LessThan.java
@@ -1,0 +1,21 @@
+package org.simple.coollection.matcher.custom;
+
+import org.simple.coollection.matcher.Matcher;
+
+public class LessThan implements Matcher {
+
+	private final Number value;
+
+	public LessThan(Number value) {
+		this.value = value;
+	}
+
+	@Override
+	public boolean match(Object matchValue) {
+		if (matchValue == null){
+			return false;
+		}
+		return ((Number) matchValue).doubleValue() < value.doubleValue();
+	}
+
+}
