+++ src/main/java/org/simple/coollection/matcher/custom/Not.java
@@ -1,0 +1,18 @@
+package org.simple.coollection.matcher.custom;
+
+import org.simple.coollection.matcher.Matcher;
+
+public class Not implements Matcher {
+
+	private final Matcher matcher;
+
+	public Not(Matcher matcher) {
+		this.matcher = matcher;
+	}
+
+	@Override
+	public boolean match(Object value) {
+		return !matcher.match(value);
+	}
+
+}
