+++ src/main/java/org/simple/coollection/query/order/Order.java
@@ -1,0 +1,5 @@
+package org.simple.coollection.query.order;
+
+public enum Order {
+	ASC, DESC
+}
