+++ src/main/java/org/simple/coollection/query/specification/Specification.java
@@ -1,0 +1,9 @@
+package org.simple.coollection.query.specification;
+
+import org.simple.coollection.query.criteria.Criteria;
+
+public interface Specification<T> {
+	
+	public boolean match(T item, Criteria<T> one, Criteria<T> other);
+	
+}
