+++ src/main/java/org/simple/coollection/query/specification/custom/AndSpecification.java
@@ -1,0 +1,13 @@
+package org.simple.coollection.query.specification.custom;
+
+import org.simple.coollection.query.criteria.Criteria;
+import org.simple.coollection.query.specification.Specification;
+
+public class AndSpecification<T> implements Specification<T> {
+
+	@Override
+	public boolean match(T item, Criteria<T> one, Criteria<T> other) {
+		return one.match(item) && other.match(item);
+	}
+
+}
